pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager app"
authors = ["shizomanya <mistic_aa@inbox.ru>"]
repository = "https://github.com/shizomanya/python-project-lvl4.git"
packages = [
  { include = "task_manager" },
]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Framework :: Django :: 5.0",
    "Framework :: Flake8",
    "Topic :: Education",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
django-bootstrap4 = "^24.3"
whitenoise = "^6.0.0"
django-extensions = "^3.2.3"
django-filter = "^24.2"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
coverage = "^7.5.3"
ipython = "^8.25.0"
pytest = "^8.2.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_paths = ["task_manager"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"